+++ src/dubai_real_estate/sql_files/tables/dld_licenced_owner_associations/create.sql
@@ -2,7 +2,7 @@ CREATE OR REPLACE TABLE
 	"{dld_database}"."{dld_table}_staging"(
 		"company_name_en" Varchar(192),
 		"company_name_ar" Nullable(Varchar(332)),
-		"phone" Nullable(Integer),
+		"phone" Nullable(Int128),
 		"email" Nullable(Varchar(76)),
 		"latitude" Nullable(Float),
 		"longitude" Nullable(Float)
