+++ src/dubai_real_estate/sql_files/tables/dld_map_requests/create.sql
@@ -1,12 +1,12 @@
 CREATE OR REPLACE TABLE
 	"{dld_database}"."{dld_table}_staging"(
-		"application_id" Integer,
-		"procedure_id" Nullable(Integer),
-		"property_type_id" Nullable(Integer),
-		"request_id" Nullable(Integer),
-		"request_source_id" Nullable(Integer),
+		"application_id" Int128,
+		"procedure_id" Nullable(Int128),
+		"property_type_id" Nullable(Int128),
+		"request_id" Nullable(Int128),
+		"request_source_id" Nullable(Int128),
 		"request_date" Nullable(Date),
-		"no_of_siteplans" Nullable(Integer)
+		"no_of_siteplans" Nullable(Int128)
 	) 
     ENGINE = MergeTree()
     PRIMARY KEY("application_id");
