+++ src/dubai_real_estate/sql_files/tables/dld_oa_service_charges/create.sql
@@ -1,18 +1,18 @@
 CREATE OR REPLACE TABLE
 	"{dld_database}"."{dld_table}_staging"(
-		"master_community_id" Nullable(Integer),
-		"project_id" Nullable(Integer),
+		"master_community_id" Nullable(Int128),
+		"project_id" Nullable(Int128),
 		"project_name" Nullable(Varchar(116)),
-		"budget_year" Integer,
-		"property_group_id" Nullable(Integer),
+		"budget_year" Int128,
+		"property_group_id" Nullable(Int128),
 		"property_group_name_en" Nullable(Varchar(200)),
 		"property_group_name_ar" Nullable(Varchar(266)),
-		"management_company_id" Nullable(Integer),
+		"management_company_id" Nullable(Int128),
 		"management_company_name_en" Nullable(Varchar(158)),
 		"management_company_name_ar" Nullable(Varchar(252)),
-		"usage_id" Nullable(Integer),
-		"service_category_id" Nullable(Integer),
-		"service_cost" Nullable(Integer)
+		"usage_id" Nullable(Int128),
+		"service_category_id" Nullable(Int128),
+		"service_cost" Nullable(Int128)
 	) 
     ENGINE = MergeTree()
     PRIMARY KEY("budget_year");
