+++ src/dubai_real_estate/sql_files/tables/dld_valuation/create.sql
@@ -1,11 +1,11 @@
 CREATE OR REPLACE TABLE
 	"{dld_database}"."{dld_table}_staging"(
-		"procedure_number" Nullable(Integer),
-		"procedure_year" Nullable(Integer),
+		"procedure_number" Nullable(Int128),
+		"procedure_year" Nullable(Int128),
 		"instance_date" Nullable(Date),
-		"area_id" Integer,
-		"property_type_id" Nullable(Integer),
-		"property_sub_type_id" Nullable(Integer),
+		"area_id" Int128,
+		"property_type_id" Nullable(Int128),
+		"property_sub_type_id" Nullable(Int128),
 		"row_status_code" Nullable(Varchar(20)),
 		"procedure_area" Nullable(Float),
 		"actual_area" Nullable(Float),
